independant example for loading a glb to point cloud
